# Conflicts:
#	ifs-data-layer/ifs-data-service/src/main/resources/db/webtest/V157_2_8__Base_webtest_data.sql
#	ifs-data-layer/ifs-data-service/src/test/resources/testdata/assessor-users.csv
#	robot-tests/IFS_acceptance_tests/tests/04__Applicant/03__Application_Form_tests/02__Finance_section/06__Academic_finances.robot
#	robot-tests/IFS_acceptance_tests/tests/04__Applicant/03__Application_Form_tests/02__Finance_section/07__Finance_summary_page.robot
#	robot-tests/IFS_acceptance_tests/tests/12__Administrator/01__Manage_users_and_applications.robot

+++ ifs-resources/src/main/java/org/innovateuk/ifs/finance/resource/BaseFinanceResource.java
@@ -154,7 +154,7 @@ public abstract class BaseFinanceResource {
     }
 
     @JsonIgnore
-    public int getGrantClaimPercentage() {
+    public BigDecimal getGrantClaimPercentage() {
         GrantClaim grantClaim = getGrantClaim();
         return grantClaim.calculateClaimPercentage(getTotal(), getTotalOtherFunding());
     }
